stable v2 + size, setRect ok

+++ app/src/main/java/com/example/kotclash/models/Map.kt
@@ -29,8 +29,9 @@ class Map()  {
         val xx = gameObject.coordinates.first.toInt()
         val yy = gameObject.coordinates.second.toInt()
 
-        for (x in xx-szx..xx+szx){
-            for (y in yy-szy..yy+szy){
+        //TODO Clean here
+        for (x in (xx-szx).toInt()..(xx+szx).toInt()){
+            for (y in (yy-szy).toInt()..(yy+szy).toInt()){
                 //to account for non-existing tiles
                 try {
                     grid[y][x].setOccupant(gameObject as Entity) //TODO we need to decide if we add objects or entities
